Move WinSports extractor to mediastream.py
